redirect is enhanced to work with buttons. auth0 domain and client moved to a separate file
