+++ frontend/src/auth0components/Auth0Addresses.ts
@@ -1,0 +1,16 @@
+//AcademicHub Application
+
+// enum Auth0Addresses{
+//     AUTH0DOMAIN = "dev-jwelisx6nixm5e0g.us.auth0.com",
+//     AUTH0CLIENTID = "GnfPQ0HkS1tLwwDbJ7rWMtynGTI6G1X4"
+// }
+// export {Addresses};
+
+
+//AcademicHubLocal Application
+
+enum Auth0Addresses{
+    AUTH0DOMAIN = "dev-jwelisx6nixm5e0g.us.auth0.com",
+    AUTH0CLIENTID = "jk5S9hWkgxu0eJwxfbRoPuwNnbzNHOIb"
+}
+export {Auth0Addresses};
