+++ frontend/src/auth0components/auth0Login.tsx
@@ -12,7 +12,11 @@ const LoginButton = () => {
         }
     }, [isAuthenticated, user]);
 
-    return <button onClick={() => loginWithRedirect()}>Log In</button>;
+    return <button onClick={() => loginWithRedirect({
+        appState: {
+            returnTo: `${window.location.origin}${window.location.pathname}`
+        }
+    })}>Log In</button>;
 };
 
 export default LoginButton;
