+++ frontend/src/auth0components/authProfile.tsx
@@ -2,6 +2,7 @@ import { useAuth0 } from "@auth0/auth0-react";
 import React, {useEffect} from "react";
 import LogoutButton from "./auth0Logout";
 import {loginService} from "../api/loginService";
+import LoginButton from "./auth0Login";
 
 const Profile = () => {
     const { user, isAuthenticated, isLoading, loginWithRedirect} = useAuth0();
@@ -11,8 +12,11 @@ const Profile = () => {
     }
 
     if (!isAuthenticated || !user) {
-        loginWithRedirect()
-        return <div>Not authenticated</div>;
+        return(
+            <div>
+                <p>You are not authenticated. Please login</p>
+                <LoginButton></LoginButton>
+            </div>);
     }
 
     return (
