+++ frontend/src/pages/ProfilePage/Profile.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import {useAuth0} from "@auth0/auth0-react";
+import LoginButton from "../../auth0components/auth0Login";
 
 function ProfilePage() {
 
@@ -13,13 +14,13 @@ function ProfilePage() {
     }
 
     if (!isAuthenticated || !user) {
-        loginWithRedirect({
-            appState: {
-                returnTo: `${window.location.origin}${window.location.pathname}`
-            }
-        })
-        return <div>Not authenticated</div>;
+        return(
+            <div>
+                <p>You are not authenticated. Please login</p>
+                <LoginButton></LoginButton>
+            </div>);
     }
+
     const handleButtonClick = () => {
         setDisplayedText(`Name: ${inputText1}<br>Last name: ${inputText2}`);
     };
